revert: move llrt to llrt_modules
