+++ llrt_core/Cargo.toml
@@ -22,6 +22,9 @@ rquickjs-core = { path = "target/patch/rquickjs-core-0.6.2" }
 llrt_modules = { path = "../llrt_modules", features = ["all"] }
 llrt_utils = { path = "../llrt_utils", features = ["all"] }
 chrono = { version = "0.4.38", default-features = false, features = ["std"] }
+quick-xml = "0.36.0"
+crc32c = { version = "0.6.8" }
+crc32fast = "1.4.2"
 phf = "0.11.2"
 rquickjs = { version = "0.6.2", features = [
     "full-async",
@@ -30,6 +33,15 @@ rquickjs = { version = "0.6.2", features = [
 ], default-features = false }
 tokio = { version = "1", features = ["full"] }
 tracing = { version = "0.1.40", features = ["log"] }
+uuid = { version = "1.10.0", default-features = false, features = [
+    "v1",
+    "v3",
+    "v4",
+    "v5",
+    "v6",
+    "v7",
+    "fast-rng",
+] }
 once_cell = "1.19.0"
 tracing-core = "0.1.32"
 webpki = "0.22.4"
@@ -37,6 +49,7 @@ async-trait = "0.1.81"
 zstd = { version = "0.13.2", default-features = false, features = [] }
 libc = "0.2.155"
 bytes = "1.6.1"
+uuid-simd = "0.8.0"
 simd-json = { version = "0.13.10", default-features = false, features = [
     "big-int-as-float",
 ] }
