+++ llrt_core/src/modules/mod.rs
@@ -2,14 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub use llrt_modules::{
-    buffer, child_process, crypto, exceptions, fs, llrt, navigator, net, os, path, perf_hooks,
-    process, zlib,
+    buffer, child_process, crypto, exceptions, fs, navigator, net, os, path, perf_hooks, process,
+    zlib,
 };
 
 pub mod console;
 pub mod encoding;
 pub mod events;
 pub mod http;
+pub mod llrt;
 pub mod module;
 pub mod url;
 pub mod util;
