+++ llrt_modules/Cargo.toml
@@ -16,7 +16,6 @@ all = [
   "events",
   "exceptions",
   "fs",
-  "llrt",
   "navigator",
   "net",
   "os",
@@ -29,11 +28,10 @@ all = [
 
 buffer = ["llrt_utils/encoding"]
 child-process = ["events", "__stream", "itoa"]
-crypto = ["crc32c", "crc32fast", "md-5", "rand", "llrt"]
+crypto = ["crc32c", "crc32fast", "md-5", "rand", "uuid"]
 events = []
 exceptions = []
 fs = ["tokio/fs", "llrt_utils/fs", "ring", "buffer", "path"]
-llrt = ["uuid", "uuid-simd", "quick-xml"]
 navigator = []
 net = ["events", "__stream"]
 os = [
@@ -73,16 +71,9 @@ flate2 = { version = "1.0.30", features = [
 ], default-features = false, optional = true }
 brotlic = { version = "0.8.2", optional = true }
 uuid = { version = "1.10.0", default-features = false, features = [
-    "v1",
-    "v3",
     "v4",
-    "v5",
-    "v6",
-    "v7",
     "fast-rng",
 ], optional = true }
-uuid-simd = { version = "0.8.0", optional = true }
-quick-xml = { version = "0.36.0", optional = true }
 rand = { version = "0.8.5", optional = true }
 crc32c = { version = "0.6.8", optional = true }
 crc32fast= { version = "1.4.2", optional = true }
