+++ types/index.d.ts
@@ -7,9 +7,6 @@
 /// <reference types="./fs.d.ts" />
 /// <reference types="./fs/promises.d.ts" />
 /// <reference types="./globals.d.ts" />
-/// <reference types="./llrt/hex.d.ts" />
-/// <reference types="./llrt/uuid.d.ts" />
-/// <reference types="./llrt/xml.d.ts" />
 /// <reference types="./navigator.d.ts" />
 /// <reference types="./net.d.ts" />
 /// <reference types="./os.d.ts" />
