--- types/llrt/hex.d.ts
@@ -1,4 +1,0 @@
-declare module "llrt:hex" {
-  export const decode: (text: string) => Uint8Array;
-  export const encode: (bytes: Uint8Array) => string;
-}
